+++ app/__init__.py
@@ -17,4 +17,5 @@ manager = Manager(app)
 manager.add_command('db', MigrateCommand)
 
 from app.controllers import default
-
+from app.controllers.login import login
+from app.controllers.usuario import usuario_controller
