+++ app/controllers/default.py
@@ -8,145 +8,13 @@ from werkzeug.security import generate_password_hash, check_password_hash
 
 
 
-# AUTENTICAÇÃO
-# *********************************************************
-
-@login_manager.user_loader
-def current_user(id):
-    return UsuarioModel.query.get(id)
-
-@app.route('/')
-@app.route('/login',methods =["GET","POST"])
-def login():
-	if request.method =="POST":
-		email = request.form['email']
-		password = request.form['password']
-		
-		usuario = UsuarioModel.query.filter_by(email=email).first()
-
-		if not usuario:
-			flash("Credênciais incorretas")
-			return redirect(url_for("login"))
-
-		if not check_password_hash(usuario.password, password):
-			flash("Senha  incorretas")
-			return redirect(url_for("login"))
-
-		login_user(usuario)
-		return redirect(url_for('listagem'))
-	
-	return render_template('login.html')
-
-
-@app.route("/logout")
-@login_required
-def logout():
-    logout_user()
-    return render_template("login.html")
-
-# *********************************************************
-
-
-
-#  USUARIO
-# *********************************************************
-
-@app.route('/cadastrarUsuario')
-@login_required
-def cadastrarUsuario():
-	return render_template('cadastroUsuario.html')
-
-@app.route('/salvar_cadastro',methods=['POST'])
-@login_required
-def salvar_cadastro():
-	nome = request.form.get('nome')
-	email = request.form.get('email')
-	password = generate_password_hash(request.form["password"])
-	endereco = request.form.get('endereco')
-	cidade = request.form.get('cidade')
-	estado = request.form.get('estado')
-	cep = request.form.get('cep')			
-	tipoUsuario = request.form.get('tipoUsuario')
-
-	usuario = UsuarioModel(nome,email,password,endereco,cidade,estado,cep,tipoUsuario)
-
-	db.session.add(usuario)
-	db.session.commit()
-
-	usuarios = UsuarioModel.query.all()
-	return render_template('listarUsuarios.html', usuarios=usuarios)
-
-@app.route('/listarUsuarios')
-@login_required
-def listarUsuarios():
-	usuarios = UsuarioModel.query.all()
-	return render_template('listarUsuarios.html', usuarios=usuarios)
-
-
-@app.route('/deletarUsuario/<int:id>')
-@login_required
-def deletarUsuario(id=0):
-	usuario = usuarios = UsuarioModel.query.filter_by(id=id).first()
-
-	return render_template('deletarUsuario.html', usuario=usuario)
-
-@app.route('/savedeletarUsuario',methods=['POST'])
-@login_required
-def savedeletarUsuario():
-	id = int(request.form.get('id'))
-
-	usuario = usuarios = UsuarioModel.query.filter_by(id=id).first()
-	
-	db.session.delete(usuario)
-	db.session.commit()
-	
-	usuarios = UsuarioModel.query.all()
-	return render_template('listarUsuarios.html', usuarios=usuarios)
-
-@app.route('/editarUsuario/<int:id>')
-@login_required
-def editarUsuario(id=0):
-	usuario = usuarios = UsuarioModel.query.filter_by(id=id).first()
-	return render_template('editarUsuario.html', usuario=usuario)
-
-
-@app.route('/saveEditarUsuario',methods=['POST'])
-@login_required
-def saveeditarUsuario():
-	id = int(request.form.get('id'))
-	nome_form = request.form.get('nome')
-	email_form = request.form.get('email')
-	endereco_form = request.form.get('endereco')
-	cidade_form = request.form.get('cidade')
-	estado_form = request.form.get('estado')
-	cep_form = request.form.get('cep')			
-	tipoUsuario_form = request.form.get('tipoUsuario')
-
-	usuarios = UsuarioModel.query.filter_by(id=id).first()
-
-	usuarios.nome = nome_form
-	usuarios.email = email_form
-	usuarios.endereco = endereco_form
-	usuarios.cidade = cidade_form
-	usuarios.estado = estado_form
-	usuarios.cep = cep_form
-	usuarios.tipoUsuario = tipoUsuario_form
-
-	db.session.commit()
-
-	usuarios = UsuarioModel.query.all()
-	return render_template('listarUsuarios.html', usuarios=usuarios)
-
-# *********************************************************
-
-
 #  PRODUTOS
 # *********************************************************
 
-@app.route('/cadastrarProduto')
-@login_required
-def cadastrarProduto():
-	return render_template('cadastroProduto.html')
+# @app.route('/cadastrarProduto')
+# @login_required
+# def cadastrarProduto():
+# 	return render_template('cadastroProduto.html')
 
 @app.route('/listagem')
 @login_required
