+++ app/templates/layout.html
@@ -1,0 +1,78 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1">
+	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
+	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
+	<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
+	<meta http-equiv='cache-control' content='no-cache'>
+	<meta http-equiv='expires' content='0'>
+	<meta http-equiv='pragma' content='no-cache'>
+	
+
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/jspdf/1.4.1/jspdf.debug.js" 
+	integrity="sha384-THVO/sM0mFD9h7dfSndI6TS0PgAGavwKvB5hAxRRvc0o9cPLohB0wb/PTA7LdUHs" 
+	crossorigin="anonymous"></script>
+</head>
+<body>
+    <nav class="navbar navbar-expand-lg navbar navbar-expand-sm bg-dark">
+		<a class="navbar-brand" href="#">Página Inicial</a>
+		<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+		  <span class="navbar-toggler-icon"></span>
+		</button>
+	  
+		<div class="collapse navbar-collapse" id="navbarSupportedContent">
+		  <ul class="navbar-nav mr-auto">
+			<li class="nav-item active">
+			  <a class="nav-link" href="#">Home <span class="sr-only">(current)</span></a>
+			</li>
+			<li class="nav-item dropdown">
+			  <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+				Existente
+			  </a>
+			  <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+				<a class="dropdown-item" href="/listagem">Listagem</a>
+			  	<a class="dropdown-item" href="/insercao">Cadastro</a>
+				<div class="dropdown-divider"></div>
+				<a class="dropdown-item" href="/graficos">Graficos</a>
+			  </div>
+			</li>
+			<li class="nav-item dropdown">
+				<a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+					Usuário
+				</a>
+				<div class="dropdown-menu" aria-labelledby="navbarDropdown">
+				  <a class="dropdown-item" href="/cadastrarUsuario">Cadastrar Usuário</a>
+				  <a class="dropdown-item" href="/listarUsuarios">Listar Usuario</a>
+				  <div class="dropdown-divider"></div>
+				  <a class="dropdown-item" href="/graficos">Graficos</a>
+				</div>
+			  </li>
+			  <li class="nav-item dropdown">
+				<a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+					Produtos
+				</a>
+				<div class="dropdown-menu" aria-labelledby="navbarDropdown">
+				  <a class="dropdown-item" href="/cadastrarProduto">Cadastrar Produto</a>
+				  <a class="dropdown-item" href="/listarProdutos">Listar Produto</a>
+				  <div class="dropdown-divider"></div>
+				  <a class="dropdown-item" href="/graficos">Graficos</a>
+				</div>
+			  </li>
+		  </ul>
+		  </ul>
+		  <form class="form-inline my-2 my-lg-0">
+			<a href="/logout" class="btn-link my-2 my-sm-0" role="button" aria-pressed="true">Sair</a>
+		  </form>
+		</div>
+	  </nav>
+	
+	<br>
+	<div class="container">
+        
+        {% block rolf %} dadas {% endblock %}
+	</div>
+</body>
+</html>
